refactor: move connection pool code to separate file
